requirements.txt
feedparser
tk
pywebview
requests
selenium